update: move NetworkUtil to network package.
