+++ app/src/main/java/com/example/armsapp/network/service/ProjectServiceImpl.kt
@@ -3,7 +3,7 @@ package com.example.armsapp.network.service
 import com.example.armsapp.network.dto.ProjectDto
 import com.example.armsapp.utils.AndroidLogger
 import com.example.armsapp.utils.Logger
-import com.example.armsapp.utils.safeApiCall
+import com.example.armsapp.network.utils.safeApiCall
 import io.ktor.client.HttpClient
 import io.ktor.client.call.body
 import io.ktor.client.request.get
