+++ app/src/main/java/com/example/armsapp/network/utils/NetworkUtil.kt
@@ -1,5 +1,6 @@
-package com.example.armsapp.utils
+package com.example.armsapp.network.utils
 
+import com.example.armsapp.utils.Logger
 import io.ktor.client.network.sockets.SocketTimeoutException
 
 suspend fun <T> safeApiCall(logger: Logger, block: suspend () -> T): Result<T> {
